+++ www/commit-hash.js
@@ -1,1 +1,1 @@
-window.__COMMIT__ = '31413379be674280b79aabea3eb449c71148c9bf';
+window.__COMMIT__ = 'd565f6029290e8a90d98367faa6e6c675063eb28';